Add Singleton design pattern

+++ src/main/java/com/example/project1/Delete.java
@@ -2,6 +2,7 @@ package com.example.project1;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
+import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.List;
 
@@ -9,20 +10,27 @@ import java.util.List;
 public abstract class Delete {
 
     public static void deleteFromDB(String passport){
-        List<UserData> userList = FetchFromDB.fetchUserData();
-        for (UserData user : userList){
-            if (user.getPassportNum().equals(passport)){
-                String deleteQuery = "DELETE FROM clients_data WHERE passport = ?";
-                try {
+        ResultSet resultSet = FetchFromDB.fetchUserData();
+        while(true){
+            try {
+                if (!resultSet.next()) break;
+
+                SingleUserData user = SingleUserData.getInstance();
+                user.setAll(resultSet.getString("first_name"),resultSet.getString("second_name"),resultSet.getString("passport"),resultSet.getString("email"),resultSet.getInt("reserved_room"),resultSet.getString("reservation_period"));
+
+                if (user.getPassportNum().equals(passport)){
+                    String deleteQuery = "DELETE FROM clients_data WHERE passport = ?";
                     Connection connection = DBConnection.connection();
                     PreparedStatement stmt = connection.prepareStatement(deleteQuery);
                     stmt.setString(1,user.getPassportNum());
                     stmt.executeUpdate();
-                } catch (SQLException e) {
-                    throw new RuntimeException(e);
                 }
+
+            } catch (SQLException e) {
+                throw new RuntimeException(e);
             }
         }
+
     }
 
     public static void deleteUser(String passport){
